Compile fix for mac/linux.
git-svn-id: file:///home/nick/svnrepo/repo/new_cyberspace/trunk@11059 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ gui_client/ObjectEditor.cpp
@@ -122,7 +122,7 @@ void ObjectEditor::setFromObject(const WorldObject& ob, int selected_mat_index_)
 		SignalBlocker blocker(this->materialComboBox);
 		this->materialComboBox->clear();
 		for(size_t i=0; i<ob.materials.size(); ++i)
-			this->materialComboBox->addItem(QtUtils::toQString("Material " + toString(i)), i);
+			this->materialComboBox->addItem(QtUtils::toQString("Material " + toString(i)), (int)i);
 
 		this->materialComboBox->setCurrentIndex(selected_mat_index);
